backend/requirements.txt
Flask==3.1.0
Flask_Cors==4.0.0
python-dotenv==1.0.1
supabase==2.13.0
